Move to Typescript

+++ db/migrate.ts
@@ -1,15 +1,16 @@
 import config from 'config'
 import { MongoClient } from 'mongodb'
-import migrations from './migrations/index.js'
-import {Migration} from './schema.js'
+import { ConnectOptions } from 'mongoose'
+import migrations from './migrations/index'
+import {Migration} from './schema'
 
-const MONGO_URL = config.get('mongoUri')
+const MONGO_URL: string = config.get('mongoUri')
 const migrationType = config.get('migrationType')
 
 export const getDb = async () => {
-    const client = await MongoClient.connect(MONGO_URL, { useUnifiedTopology: true });
+    const client = await MongoClient.connect(MONGO_URL, { useUnifiedTopology: true } as ConnectOptions)
     return client.db()
-};
+}
 
 export const migrate = async () => {
     const db = await getDb()
@@ -39,6 +40,10 @@ export const migrate = async () => {
             } else {
                 await func.down(db)
                 const declinedMigration = await Migration.findOne({name})
+                if (!declinedMigration) {
+                    console.log(`Migration ${name} doesn't provided yet`)
+                    return
+                }
                 await declinedMigration.deleteOne()
                 console.log(`Migration ${name} cancel successfully`)
             }
