+++ routes/order.routes.ts
@@ -1,8 +1,8 @@
-import {Router} from 'express'
+import express, {Router} from 'express'
 import {validationResult} from "express-validator"
-import {Order} from '../models/Order.js'
-import {Product} from '../models/Product.ts'
-// import {orderValidators} from '../utils/validators'
+import {Order} from '../models/Order'
+import {Product} from '../models/Product'
+import {orderValidators} from '../utils/validators'
 
 
 export const router = Router()
@@ -29,8 +29,9 @@ router.get('/:id', async (req, res) => {
     }
 })
 
-router.post('/', [orderValidators],
-    async (req, res) => {
+router.post('/',
+    orderValidators,
+    async (req: express.Request, res: express.Response) => {
         try {
             const errors = validationResult(req)
             if (!errors.isEmpty()) {
@@ -39,19 +40,24 @@ router.post('/', [orderValidators],
 
             const {user, product, qty} = req.body
             const productInstance = await Product.findById(product)
-            const sum = productInstance.price * qty
+            if (productInstance?.price) {
+                const sum = productInstance.price * qty
 
-            const order = new Order({user, product, qty, sum})
-            await order.save()
+                const order = new Order({user, product, qty, sum})
+                await order.save()
 
-            res.status(201).json('Order added')
+                res.status(201).json('Order added')
+            } else {
+                res.status(404).json({message: 'Product does not exist'})
+            }
         } catch (e) {
             res.status(404).json('Bad request')
         }
     })
 
-router.patch('/:id', [orderValidators],
-    async (req, res) => {
+router.patch('/:id',
+    orderValidators,
+    async (req: express.Request, res: express.Response) => {
         try {
             const errors = validationResult(req)
             const params = req.body
@@ -64,6 +70,10 @@ router.patch('/:id', [orderValidators],
             }
             const order = await Order.findById(req.params.id)
 
+            if (!order) {
+                res.status(404).json({message: 'Order does not exist'})
+                return
+            }
             if (params.product || params.qty) {
                 let productInstance, qty
                 if (params.product) {
@@ -75,6 +85,10 @@ router.patch('/:id', [orderValidators],
                     qty = params.qty
                 } else {
                     qty = order.qty
+                }
+                if (!productInstance || ! productInstance.price) {
+                    res.status(404).json({message: 'Product in order does not exist'})
+                    return
                 }
                 params.sum = productInstance.price * qty
             }
@@ -91,9 +105,13 @@ router.patch('/:id', [orderValidators],
 router.delete('/:id', async (req, res) => {
     try {
         const order = await Order.findById(req.params.id)
+        if (!order) {
+            res.status(404).json({message: 'Order does not exist'})
+            return
+        }
         await order.deleteOne()
 
-        res.status(200).json('Product deleted')
+        res.status(200).json('Order deleted')
     } catch (e) {
         res.status(404).json({message: 'Bad request'})
     }
